garmin_upload/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
paho-mqtt==1.5.1
pkg-resources==0.0.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4

